modules/revspy/requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
httpx>=0.27.0
pydantic>=2.7.0
python-multipart>=0.0.6
aiofiles>=23.2.1
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
reportlab>=4.0.0
